@@ -1,5 +1,3 @@
-from typing import Union
-
 import pytest
 
 from template_python.examples import fizzbuzz
@@ -9,5 +7,5 @@ from template_python.examples import fizzbuzz
     "input_item,output",
     [(1, 1), (2, 2), (3, "Fizz"), (4, 4), (5, "Buzz"), (15, "FizzBuzz")],
 )
-def test_fizzbuzz(input_item: int, output: Union[int, str]) -> None:
+def test_fizzbuzz(input_item: int, output: int | str) -> None:
     assert fizzbuzz.fizzbuzz_list()[input_item] == output
